Resolves #6

+++ Debian/stretch64/scripts/cleanup.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -e
+
 # Clean up
 apt-get -y remove linux-headers-$(uname -r) build-essential
 apt-get -y autoremove
@@ -11,10 +13,16 @@ rm /var/lib/dhcp/*
 
 # Make sure Udev doesn't block our network
 echo "cleaning up udev rules"
-rm -rf /etc/udev/rules.d/70-persistent-net.rules
+if [ -d /etc/udev/rules.d/70-persistent-net.rules ]; then
+  rm -rf /etc/udev/rules.d/70-persistent-net.rules
+fi
 # mkdir /etc/udev/rules.d/70-persistent-net.rules
-rm -rf /dev/.udev/
-rm /lib/udev/rules.d/75-persistent-net-generator.rules
+if [ -d /dev/.udev/ ]; then
+  rm -rf /dev/.udev/
+fi
+if [ -f /lib/udev/rules.d/75-persistent-net-generator.rules ]; then
+  rm /lib/udev/rules.d/75-persistent-net-generator.rules
+fi
 
 echo "Adding a 2 sec delay to the interface up, to make the dhclient happy"
 echo "pre-up sleep 2" >> /etc/network/interfaces
